Try to always detect the local IP address.
Before starting the local server, try to detect the IP address
and bind the server onto the detected address. Also, use the same
address in the medic.json configuration so that test results gets
posted to the correct URL.

This change allows posting results from a Windows Phone emulator
onto a server running on the host machine. In that case, posting
to localhost doesn't work since the emulator has its own networking
configuration.

This change also removes the dependency to the localtunnel module,
which isn't used anywhere at the moment.